Approved microplan integration
