+++ health/micro-ui/web/micro-ui-internals/packages/modules/campaign-manager/src/pages/employee/index.js
@@ -14,7 +14,7 @@ import ViewHierarchy from "./ViewHierarchy";
 import ViewChecklist from "./ViewChecklist";
 import UpdateChecklist from "./UpdateChecklist";
 import BoundaryHome from "./BoundaryHome";
-import MyMicroplans from "./MyMicroplans";
+import ApprovedMicroplans from "./ApprovedMicroplans";
 import DummyLoaderScreen from "../../components/fetchFromMicroplan";
 /**
  * The CampaignBreadCrumb function generates breadcrumb navigation for a campaign setup page in a React
@@ -153,7 +153,7 @@ const App = ({ path, BOUNDARY_HIERARCHY_TYPE, hierarchyData }) => {
           <PrivateRoute path={`${path}/boundary/view-all-hierarchy`} component={()=> <ViewBoundary />} />
           <PrivateRoute path={`${path}/boundary/data`} component={()=> <ViewHierarchy />} />
           <PrivateRoute path={`${path}/update-campaign`} component={() => <UpdateCampaign />} />
-          <PrivateRoute path={`${path}/setup-from-microplan`} component={() => <MyMicroplans />} />
+          <PrivateRoute path={`${path}/setup-from-microplan`} component={() => <ApprovedMicroplans />} />
         </AppContainer>
       </Switch>
     </React.Fragment>
